refactor: Moved appearance settigns and event handling from main.py to frame_window_setup.py,
run_server() from main.py to server_setup.py
