+++ ui/frame_window_setup.py
@@ -2,6 +2,9 @@ import tkinter as tk
 import customtkinter as ctk
 from tkinter import ttk
 
+ctk.set_appearance_mode("System")  # Modes: "System" (standard), "Dark", "Light"
+ctk.set_default_color_theme("blue")  # Themes: "blue" (standard), "green", "dark-blue"
+
 # Create the Tkinter root
 root = ctk.CTk()
 root.title("Opera Access 1.0")
@@ -108,6 +111,7 @@ navigation_frame2.grid(row=8, column=2,rowspan=1, padx=0, pady=0, sticky="w")
 navigation_frame3 = ctk.CTkFrame(root,fg_color="transparent", width=500, height=200, corner_radius=4, border_width=0)
 navigation_frame3.grid(row=8, column=1,rowspan=1, padx=0, pady=0, sticky="e")
 
+
 # Create a 'Go' button to jump to the specified line number
 go_button = ctk.CTkButton(navigation_frame2,fg_color="transparent", text_color=("gray10", "#DCE4EE"),border_width=2, text="Go", command=jump_to_line)
 go_button.grid(row=0, column=0, padx=10, pady=10, sticky="w")
@@ -117,3 +121,9 @@ previous_button.grid(row=1, column=0, padx=10, pady=10, sticky="e")
 
 next_button = ctk.CTkButton(navigation_frame2,fg_color="transparent", text_color=("gray10", "#DCE4EE"),border_width=2, text="Next", command=next_line)
 next_button.grid(row=1, column=0, padx=10, pady=10, sticky="w")
+
+def on_key_press(event):
+    if event.keysym == 'Right':
+        next_line()
+    elif event.keysym == 'Left':
+        previous_line()
